Add utility for tag seasonality
