+++ src/NavTagMenu.jsx
@@ -1,8 +1,8 @@
 // src/NavTagMenu.jsx
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { RRule } from 'rrule';
 import { supabase } from './supabaseClient';
+import { isTagActive } from './utils/tagUtils';
 
 const pillStyles = [
   'bg-green-100 text-indigo-800',
@@ -19,46 +19,15 @@ export default function NavTagMenu() {
   const [tags, setTags] = useState([]);
   const [seasonalTags, setSeasonalTags] = useState([]);
 
-  function parseLocalYMD(str) {
-    if (!str) return null;
-    const [y, m, d] = str.split('-').map(Number);
-    return new Date(y, m - 1, d);
-  }
-
-  function isTagActive(tag) {
-    const today = new Date();
-    today.setHours(0, 0, 0, 0);
-    if (tag.rrule) {
-      try {
-        const opts = RRule.parseString(tag.rrule);
-        if (tag.season_start) opts.dtstart = parseLocalYMD(tag.season_start);
-        const rule = new RRule(opts);
-        const searchStart = new Date(today);
-        searchStart.setDate(searchStart.getDate() - 8);
-        const next = rule.after(searchStart, true);
-        if (!next) return false;
-        const start = new Date(next);
-        start.setDate(start.getDate() - 7);
-        const end = new Date(next);
-        end.setDate(end.getDate() + 1);
-        return today >= start && today < end;
-      } catch {
-        return false;
-      }
-    }
-    if (tag.season_start && tag.season_end) {
-      const start = parseLocalYMD(tag.season_start);
-      const end = parseLocalYMD(tag.season_end);
-      return start && end && today >= start && today <= end;
-    }
-    return true;
-  }
-
   useEffect(() => {
     (async () => {
       try {
         const [allRes, seasonRes] = await Promise.all([
-          supabase.from('tags').select('name, slug').order('name'),
+          supabase
+            .from('tags')
+            .select('name, slug')
+            .eq('is_seasonal', false)
+            .order('name'),
           supabase
             .from('tags')
             .select('name, slug, rrule, season_start, season_end')
